+++ azext_iot/operations/hub.py
@@ -1116,10 +1116,11 @@ def iot_device_upload_file(cmd, device_id, file_path, content_type, hub_name=Non
 # pylint: disable=too-many-locals
 def iot_hub_monitor_events(cmd, hub_name=None, device_id=None, consumer_group='$Default', timeout=300,
                            enqueued_time=None, resource_group_name=None, yes=False, properties=None, repair=False,
-                           login=None, content_type=None):
+                           login=None, content_type=None, device_query=None):
     import importlib
     from azext_iot.common.deps import ensure_uamqp
     from azext_iot.common.utility import validate_min_python_version
+
     validate_min_python_version(3, 5)
 
     if timeout < 0:
@@ -1141,6 +1142,13 @@ def iot_hub_monitor_events(cmd, hub_name=None, device_id=None, consumer_group='$
     if not enqueued_time:
         enqueued_time = calculate_millisec_since_unix_epoch_utc()
 
+    device_ids = {}
+    if device_query:
+        devices_result = iot_query(cmd, device_query, hub_name, None, resource_group_name, login=login)
+        if devices_result:
+            for device_result in devices_result:
+                device_ids[device_result['deviceId']] = True
+
     target = get_iot_hub_connection_string(cmd, hub_name, resource_group_name, include_events=True, login=login)
     events3.executor(target,
                      consumer_group=consumer_group,
@@ -1149,7 +1157,8 @@ def iot_hub_monitor_events(cmd, hub_name=None, device_id=None, consumer_group='$
                      timeout=timeout,
                      device_id=device_id,
                      output=output,
-                     content_type=content_type)
+                     content_type=content_type,
+                     devices=device_ids)
 
 def iot_hub_monitor_feedback(cmd, hub_name=None, device_id=None, yes=False,
                              wait_on_id=None, repair=False, resource_group_name=None, login=None):